fix: eliminating dupe code
